perf(types): move types out of api module
