+++ src/api/index.ts
@@ -1,11 +1,8 @@
 import { API, ExResult } from "./api";
 import { autoAPI } from "./auto";
 import { ShadowAPI } from "./shadow";
-import { IEthereumHeaderThing, IEthereumHeaderThingWithProof } from "./types";
 
 export {
     API, autoAPI,
     ExResult, ShadowAPI,
-    IEthereumHeaderThing,
-    IEthereumHeaderThingWithProof,
 }
