+++ src/listener/cache.ts
@@ -1,7 +1,7 @@
 import { Config } from "../util";
 import path from "path";
 import fs from "fs";
-import { IEthereumHeaderThingWithProof } from "../api/types";
+import { IEthereumHeaderThingWithProof } from "../types";
 
 const cache = path.resolve((new Config()).path.root, "cache/blocks");
